fix(protocols): silence gcc -Wsign-compare and -Wformat-truncation warnings
GCC (real, not Apple clang) flagged three compile warnings under the
tsan/asan Makefile targets:

- database.c: ?: operand changed signedness (ptrdiff_t vs size_t); cast
  the pointer difference to size_t.
- tcp.c / udp.c: port_str[8] too small for worst-case "%d" of an int
  (up to 11 chars); bump to [16].

Apple clang did not emit these. Verified 0 warnings under both gcc-16
and clang; tsan/asan checks still pass.

+++ src/protocols/database.c
@@ -109,7 +109,7 @@ int database_parse_connection_string(const char* connection_string, char* host, 
         *port = atoi(colon + 1);
     } else {
         // Host without port
-        size_t host_len = slash ? (slash - url_part) : strlen(url_part);
+        size_t host_len = slash ? (size_t)(slash - url_part) : strlen(url_part);
         strncpy(host, url_part, host_len);
         host[host_len] = '\0';
